requirements.txt
apache-airflow==2.7.1
snowflake-connector-python[pandas]==3.6.0
dbt-core==1.7.4
dbt-snowflake==1.7.1
pandas==2.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.27
apache-airflow-providers-snowflake==4.9.0
apache-airflow-providers-dbt-cloud==2.2.0
apache-airflow-providers-postgres==5.8.0